+++ examples/injection_examples/binary_neutron_star_example.py
@@ -1,4 +1,4 @@
-#!/bin/python
+#!/usr/bin/env python
 """
 Tutorial to demonstrate running parameter estimation on a binary neutron star
 system taking into account tidal deformabilities.