+++ examples/injection_examples/calibration_example.py
@@ -1,4 +1,4 @@
-#!/bin/python
+#!/usr/bin/env python
 """
 Tutorial to demonstrate running parameter estimation with calibration
 uncertainties included.