+++ examples/open_data_examples/GW150914_minimal.py
@@ -1,4 +1,4 @@
-#!/bin/python
+#!/usr/bin/env python
 """
 Tutorial to demonstrate the minimum number of steps required to run parameter
 stimation on GW150914 using open data.